커밋
커밋

+++ jspProject/WebContent/views/notice/noticeDetailView.jsp
@@ -3,7 +3,7 @@
     pageEncoding="UTF-8"%>
 <%
 	Notice n = (Notice)request.getAttribute("notice");
-	// 글번호, 제목, 내용, 작성자 아이디, 작성일
+	// 글번호, 제목, 내용, 작성자아이디, 작성일 
 %>
 <!DOCTYPE html>
 <html>
@@ -20,18 +20,14 @@
         margin-top: 50px;
     }
 </style>
-
 </head>
 <body>
-    <%@ include file = "../common/menubar.jsp" %>
-
+    <%@ include file="../common/menubar.jsp" %>
     <div class="outer" align="center">
         <br>
         <h2 align="center">공지사항 상세보기</h2>
         <br>
-
         <table id="detail-area" border="1">
-            <!-- (tr>th+td+th+td)*3 -->
             <tr>
                 <th width="70">제목</th>
                 <td colspan="3" width="430"><%= n.getNoticeTitle() %></td>
@@ -49,18 +45,18 @@
                 </td>
             </tr>
         </table>
+       
         <br><br>
 
         <div>
             <a href="<%= contextPath %>/list.no" class="btn btn-sm btn-secondary">목록가기</a>
-
-            <!-- 현재 로그인한 사용자가 해당 글을 쓴 본인일 경우 -->
+            
+            <!-- 현재 로그인 한 사용자가 해당 글을 쓴 본인일 경우 -->
             <% if(loginUser != null && n.getNoticeWriter().equals(loginUser.getUserId())) { %>
-                <a href="<%= contextPath %>/updateForm.no?num=<%= n.getNoticeNo() %>" class="btn btn-sm btn-warning">수정하기</a>
-                <a href="<%= contextPath %>/delete.no?num=<%= n.getNoticeNo() %>" class="btn btn-sm btn-danger">삭제하기</a>
-            <% } %>
+	            <a href="<%= contextPath %>/updateForm.no?num=<%= n.getNoticeNo() %>" class="btn btn-sm btn-warning">수정하기</a>
+	            <a href="<%= contextPath %>/delete.no?num=<%= n.getNoticeNo() %>" class="btn btn-sm btn-danger" onclick="confirm('삭제하시겠습니까?');">삭제하기</a>
+	            <% } %>
         </div>
-       
     </div>
 </body>
 </html>
